requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0
pydantic==2.5.0
python-dateutil==2.8.2
pendulum==2.1.2
rapidfuzz==3.5.2
openai==1.3.7
loguru==0.7.2
pyyaml==6.0.1
click==8.1.7
python-dotenv==1.0.0
requests==2.31.0
